src/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pydeck>=0.9.1
streamlit>=1.37.0
statsmodels>=0.14.0
pydantic>=2.7.0
pyyaml>=6.0
h3>=3.7.6
